back out all restore listener changes
